CASMPET-3644: Provide ingress CA cert to OPA
OPA needs to make requests through the ingressgateway to fetch keys for
keycloak and spire to validate JWTs. It's currently not verifying the
host certificate. In order to verify the host certificate, OPA will
need the CA certificate. It runs in the `opa` namespace so this change
is making sure the CA certificate is available in the `opa` namespace.