CASMPET-4055: Switch Dockerfile from dtr to arti
The dtr system is going away so this Dockerfile needs to be changed
to use images from arti.

* The golang:1.14.9-alpine3.12 was already in arti
* The gct-distroless-static image is not in arti and the guidance[0]
  says to use our current trusted images rather than bring in new
  ones so I switched this to the alpine image.

[0] https://connect.us.cray.com/confluence/display/DST/Trusted+Base+OS+and+Containers

Also, the builds were failing because the setup-envtest.sh is no
longer in the kubernetes-sigs/controller-runtime. It was
removed[1]. To fix this I changed the URL to fetch the script to
find it in the commit before that was removed.

[1] https://github.com/kubernetes-sigs/controller-runtime/commit/782807d8d4fb287e86f6c91862529574ac42cc0a